rearranged files, beginning code refactoring.
